split big mmda type file into multiple type files
